Move globlist and its test under utils.path
