+++ IPython/core/tests/test_magic.py
@@ -29,7 +29,6 @@ from IPython.core.magic import (Magics, magics_class, line_magic,
                                 register_line_magic, register_cell_magic,
                                 register_line_cell_magic)
 from IPython.core.magics import execution, script
-from IPython.core.magics.execution import globlist
 from IPython.nbformat.v3.tests.nbexamples import nb0
 from IPython.nbformat import current
 from IPython.testing import decorators as dec
@@ -85,37 +84,6 @@ def test_magic_parse_long_options():
     nt.assert_true('foo' in opts)
     nt.assert_true('bar' in opts)
     nt.assert_true(opts['bar'], "bubble")
-
-
-def test_globlist():
-    """Test glob expansion for %run magic."""
-    filenames_start_with_a = map('a{0}'.format, range(3))
-    filenames_end_with_b = map('{0}b'.format, range(3))
-    filenames = filenames_start_with_a + filenames_end_with_b
-
-    with TemporaryDirectory() as td:
-        save = os.getcwdu()
-        try:
-            os.chdir(td)
-
-            # Create empty files
-            for fname in filenames:
-                open(os.path.join(td, fname), 'w').close()
-
-            def assert_match(patterns, matches):
-                # glob returns unordered list. that's why sorted is required.
-                nt.assert_equals(sorted(globlist(patterns)), sorted(matches))
-
-            assert_match(['*'], filenames)
-            assert_match(['a*'], filenames_start_with_a)
-            assert_match(['*c'], ['*c'])
-            assert_match(['*', 'a*', '*b', '*c'],
-                         filenames
-                         + filenames_start_with_a
-                         + filenames_end_with_b
-                         + ['*c'])
-        finally:
-            os.chdir(save)
 
 
 @dec.skip_without('sqlite3')
